requirements.txt
Authlib==1.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
cryptography==39.0.1
decorator==5.1.1
idna==3.4
pycparser==2.21
requests==2.28.2
tqdm==4.64.1
urllib3==1.26.14
validators==0.19.0
weaviate-client==3.11.0
